@@ -1,23 +1,22 @@
 import 'package:flutter/material.dart';
 import 'package:mashmart/views/account/welcome_screen.dart';
-import 'package:mashmart/views/homepage.dart';
 
-void main(){
+void main() {
   runApp(MyApp());
 }
 
 class MyApp extends StatelessWidget {
+  const MyApp({super.key});
+
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
       debugShowCheckedModeBanner: false,
       theme: ThemeData(
-        primaryColor: Colors.teal.shade700,
-        visualDensity: VisualDensity.adaptivePlatformDensity,
+        useMaterial3: true,
+        colorScheme: ColorScheme.fromSeed(seedColor: Colors.teal.shade700),
       ),
-
       home: WelcomeScreen(),
-
     );
   }
 }
