@@ -4,6 +4,8 @@ import 'package:mashmart/views/account/signup.dart';
 import 'package:mashmart/views/homepage.dart';
 
 class WelcomeScreen extends StatelessWidget {
+  const WelcomeScreen({super.key});
+
   @override
   Widget build(BuildContext context) {
     final signInButton = Material(
@@ -11,7 +13,7 @@ class WelcomeScreen extends StatelessWidget {
       color: Colors.teal.shade700,
       child: MaterialButton(
         minWidth: 150,
-        padding: EdgeInsets.fromLTRB(20.0, 15.0, 20.0, 15.0),
+        padding: const EdgeInsets.fromLTRB(20.0, 15.0, 20.0, 15.0),
         onPressed: () {
           print("SignIn");
           Navigator.of(context).push(MaterialPageRoute(builder: (_) {
@@ -20,7 +22,7 @@ class WelcomeScreen extends StatelessWidget {
         },
         child: Text("Sign In",
             textAlign: TextAlign.center,
-            style: TextStyle(fontFamily: 'Montserrat', fontSize: 20.0)
+            style: const TextStyle(fontFamily: 'Montserrat', fontSize: 20.0)
                 .copyWith(color: Colors.white, fontWeight: FontWeight.bold)),
       ),
     );
@@ -30,7 +32,7 @@ class WelcomeScreen extends StatelessWidget {
       color: Colors.grey.shade800,
       child: MaterialButton(
         minWidth: 150,
-        padding: EdgeInsets.fromLTRB(20.0, 15.0, 20.0, 15.0),
+        padding: const EdgeInsets.fromLTRB(20.0, 15.0, 20.0, 15.0),
         onPressed: () {
           print("SignUp");
           Navigator.of(context).push(MaterialPageRoute(builder: (_) {
@@ -39,7 +41,7 @@ class WelcomeScreen extends StatelessWidget {
         },
         child: Text("Sign Up",
             textAlign: TextAlign.center,
-            style: TextStyle(fontFamily: 'Montserrat', fontSize: 20.0)
+            style: const TextStyle(fontFamily: 'Montserrat', fontSize: 20.0)
                 .copyWith(color: Colors.white, fontWeight: FontWeight.bold)),
       ),
     );
@@ -51,16 +53,16 @@ class WelcomeScreen extends StatelessWidget {
           child: Column(
             children: [
               Container(
-                margin: EdgeInsets.all(20),
+                margin: const EdgeInsets.all(20),
                 alignment: Alignment.bottomCenter,
-                child: Text("Welcome to Mashmart",
+                child: const Text("Welcome to Mashmart",
                     style: TextStyle(
                       fontSize: 24,
                       fontWeight: FontWeight.bold,
                     )),
               ),
               Image.asset('assets/images/account_welcome_screen.png'),
-              SizedBox(height: 20),
+              const SizedBox(height: 20),
               Row(
                 mainAxisAlignment: MainAxisAlignment.center,
                 children: [
@@ -71,9 +73,9 @@ class WelcomeScreen extends StatelessWidget {
                   ),
                 ],
               ),
-              SizedBox(height: 30),
+              const SizedBox(height: 30),
               signInButton,
-              SizedBox(height: 10),
+              const SizedBox(height: 10),
               signUpButton,
             ],
           ),
